refactor texture sampling test and others
