+++ TestKit/CMakeLists.txt
@@ -1,9 +1,17 @@
 lyra_sample(testkit)
 find_package(stb REQUIRED)
 find_package(doctest REQUIRED)
-target_sources(lyra-testkit PRIVATE main.cpp helper.cpp)
+target_sources(lyra-testkit PRIVATE
+    main.cpp
+    ./common/helper.cpp
+    ./common/uniform.cpp
+    ./common/geometry.cpp
+    ./common/render.cpp
+    ./common/texture.cpp
+    ./common/app.cpp
+)
 target_link_libraries(lyra-testkit PRIVATE stb doctest::doctest)
-target_include_directories(lyra-testkit PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(lyra-testkit PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/common)
 
 # test cases
 # add_subdirectory(depth_test)
