+++ TestKit/common/helper.h
@@ -1,0 +1,19 @@
+#ifndef LYRA_TESTLIB_HELPER_H
+#define LYRA_TESTLIB_HELPER_H
+
+#define DOCTEST_CONFIG_NO_EXCEPTIONS
+#include <doctest/doctest.h>
+
+#include <cmath>
+#include <limits>
+#include <iostream>
+
+#include "./app.h"
+#include "./camera.h"
+#include "./vertex.h"
+#include "./uniform.h"
+#include "./geometry.h"
+#include "./render.h"
+#include "./texture.h"
+
+#endif // LYRA_TESTLIB_HELPER_H
