+++ TestKit/common/linmath.h
@@ -1,0 +1,11 @@
+#ifndef LYRA_TESTLIB_HELPER_MATH_H
+#define LYRA_TESTLIB_HELPER_MATH_H
+
+#define GLM_FORCE_RADIANS
+#define GLM_ENABLE_EXPERIMENTAL
+#define GLM_FORCE_DEPTH_ZERO_TO_ONE
+#include <glm/glm.hpp>
+#include <glm/gtx/string_cast.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+
+#endif // LYRA_TESTLIB_HELPER_MATH_H
