+++ TestKit/texture_sampling/main.cpp
@@ -58,6 +58,7 @@ struct TextureSamplingApp : public TestApp
     explicit TextureSamplingApp(const TestAppDescriptor& desc) : TestApp(desc)
     {
         setup_buffers();
+        setup_sampler();
         setup_texture();
         setup_pipeline();
     }
@@ -73,35 +74,9 @@ struct TextureSamplingApp : public TestApp
         uniform  = Uniform::create(fovy, aspect, glm::vec3(0.0, 0.0, 3.0));
     }
 
-    void setup_texture()
+    void setup_sampler()
     {
         auto& device = RHI::get_current_device();
-
-        uint width  = 4;
-        uint height = 4;
-
-        staging = execute([&] {
-            auto desc  = GPUBufferDescriptor{};
-            desc.size  = width * height * sizeof(uint32_t);
-            desc.usage = GPUBufferUsage::COPY_SRC | GPUBufferUsage::MAP_WRITE;
-            desc.label = "staging";
-            return device.create_buffer(desc);
-        });
-
-        texture = execute([&] {
-            auto desc            = GPUTextureDescriptor{};
-            desc.size.width      = width;
-            desc.size.height     = width;
-            desc.size.depth      = 1;
-            desc.array_layers    = 1;
-            desc.dimension       = GPUTextureDimension::x2D;
-            desc.format          = GPUTextureFormat::RGBA8UNORM;
-            desc.mip_level_count = 1;
-            desc.sample_count    = 1;
-            desc.usage           = GPUTextureUsage::COPY_DST | GPUTextureUsage::TEXTURE_BINDING;
-            desc.label           = "texture";
-            return device.create_texture(desc);
-        });
 
         sampler = execute([&] {
             auto desc           = GPUSamplerDescriptor{};
@@ -119,51 +94,16 @@ struct TextureSamplingApp : public TestApp
             desc.label          = "sampler";
             return device.create_sampler(desc);
         });
+    }
 
-        texview = texture.create_view();
+    void setup_texture()
+    {
+        auto tex = SimpleTexture2D::gradient(4, 4);
+        tex.upload();
 
-        staging.map(GPUMapMode::WRITE);
-        auto content = staging.get_mapped_range<uint>();
-        for (uint y = 0; y < height; y++) {
-            for (uint x = 0; x < width; x++) {
-                uint index        = y * width + x;
-                uint red          = uint(float(y) / float(height) * 255.0) & 0xFFu;
-                uint green        = uint(float(x) / float(width) * 255.0) & 0xFFu;
-                uint color        = (red << 0) | (green << 8) | 0xFF000000u;
-                content.at(index) = color;
-            }
-        }
-        staging.unmap();
-
-        auto command = execute([&]() {
-            auto desc  = GPUCommandBufferDescriptor{};
-            desc.queue = GPUQueueType::DEFAULT;
-            return device.create_command_buffer(desc);
-        });
-
-        auto copy_src           = GPUTexelCopyBufferInfo{};
-        copy_src.buffer         = staging;
-        copy_src.bytes_per_row  = 0;
-        copy_src.offset         = 0;
-        copy_src.rows_per_image = height;
-
-        auto copy_dst      = GPUTexelCopyTextureInfo{};
-        copy_dst.texture   = texture;
-        copy_dst.aspect    = GPUTextureAspect::COLOR;
-        copy_dst.mip_level = 0;
-        copy_dst.origin    = {0, 0, 0};
-
-        auto copy_ext   = GPUExtent3D{};
-        copy_ext.width  = width;
-        copy_ext.height = height;
-        copy_ext.depth  = 1;
-
-        command.resource_barrier(state_transition(texture, undefined_state(), copy_dst_state()));
-        command.copy_buffer_to_texture(copy_src, copy_dst, copy_ext);
-        command.resource_barrier(state_transition(texture, copy_dst_state(), shader_resource_state(GPUBarrierSync::PIXEL_SHADING)));
-        command.submit();
-
-        device.wait();
+        staging = tex.buffer;
+        texture = tex.texture;
+        texview = tex.view;
     }
 
     void setup_pipeline()
@@ -364,32 +304,32 @@ struct TextureSamplingApp : public TestApp
     }
 };
 
-TEST_CASE("rhi::vulkan::texture_sampling" * doctest::description("Rendering a textured triangle with the most basic graphics pipeline."))
-{
-    TestAppDescriptor desc{};
-    desc.name           = "vulkan";
-    desc.window         = false;
-    desc.backend        = RHIBackend::VULKAN;
-    desc.width          = 640;
-    desc.height         = 480;
-    desc.rhi_flags      = RHIFlag::DEBUG | RHIFlag::VALIDATION;
-    desc.compile_target = CompileTarget::SPIRV;
-    desc.compile_flags  = CompileFlag::DEBUG;
-    TextureSamplingApp(desc).run();
-}
-
-// #ifdef WIN32
-// TEST_CASE("rhi::d3d12::texture_sampling" * doctest::description("Rendering a textured triangle with the most basic graphics pipeline."))
+// TEST_CASE("rhi::vulkan::texture_sampling" * doctest::description("Rendering a textured triangle with the most basic graphics pipeline."))
 // {
 //     TestAppDescriptor desc{};
-//     desc.name           = "d3d12";
+//     desc.name           = "vulkan";
 //     desc.window         = false;
-//     desc.backend        = RHIBackend::D3D12;
+//     desc.backend        = RHIBackend::VULKAN;
 //     desc.width          = 640;
 //     desc.height         = 480;
 //     desc.rhi_flags      = RHIFlag::DEBUG | RHIFlag::VALIDATION;
-//     desc.compile_target = CompileTarget::DXIL;
+//     desc.compile_target = CompileTarget::SPIRV;
 //     desc.compile_flags  = CompileFlag::DEBUG;
 //     TextureSamplingApp(desc).run();
 // }
-// #endif
+
+#ifdef WIN32
+TEST_CASE("rhi::d3d12::texture_sampling" * doctest::description("Rendering a textured triangle with the most basic graphics pipeline."))
+{
+    TestAppDescriptor desc{};
+    desc.name           = "d3d12";
+    desc.window         = true;
+    desc.backend        = RHIBackend::D3D12;
+    desc.width          = 640;
+    desc.height         = 480;
+    desc.rhi_flags      = RHIFlag::DEBUG | RHIFlag::VALIDATION;
+    desc.compile_target = CompileTarget::DXIL;
+    desc.compile_flags  = CompileFlag::DEBUG;
+    TextureSamplingApp(desc).run();
+}
+#endif
